pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netai"
version = "1.0.0"
description = "NetAI — Open-source distributed AI training & inference. P2P resource pooling, pipeline-parallel inference for models of any size (3B to 700B+), CPU/GPU training, and community-driven model voting."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "aiohttp>=3.9",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "pydantic>=2.5",
    "httpx>=0.25",
    "PyYAML>=6.0",
    "psutil>=5.9",
    "cryptography>=41.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
gpu = ["torch>=2.1"]
dev = ["pytest>=7.4", "pytest-asyncio>=0.21", "pytest-cov>=4.1"]

[project.scripts]
netai = "netai.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"